requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.20.0
numpy>=1.24.0
opencv-python>=4.7.0
pillow>=9.5.0
scipy>=1.10.0
scikit-image>=0.20.0
tqdm>=4.65.0

# Web server
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# 3D processing
trimesh>=3.22.0
# open3d>=0.17.0 (optional, install manually if needed)
pyrender>=0.1.45

# Optional - for visualization
matplotlib>=3.7.0
ipywidgets>=8.0.0

# Optional - CUDA acceleration
# cupy>=12.0.0; platform_system != "Darwin" or platform_machine != "arm64"

# Vector database
qdrant-client>=1.6.0
pinecone-client>=2.2.1

# Huggingface integration
huggingface_hub>=0.16.0
safetensors>=0.3.2
git+https://github.com/huggingface/diffusers.git

# Environment variables
python-dotenv>=1.0.0

# For mesh visualization
pygltflib>=1.16.0

# For cloning repositories
gitpython>=3.1.30
